Disable only the ReadyRead callbacks for FileTransfer uploads. Previously, we disabled all callbacks which caused the upload socket to not be closed properly when the upload completed.
References #8519.  This solves the issue in the ticket, in a
different way than the patch attached to the ticket.

After a few days, if there are no regressions, I'll commit
this to -fixes as well and close the ticket.

This bumps the binary API version since libmythdb is modified,
so make clean, recompiled plugins, whatever you normally do in
this case.



git-svn-id: http://svn.mythtv.org/svn/trunk@24994 7dbf422c-18fa-0310-86e9-fd20926502f2