Miscellaneous grammar/consistency/typo fixes for setup/settings.
Fixes include:

{{{
* Myth -> MythTV
* Changing all occurences of 2 spaces after a period to one space (we had a mix of both).
* Consistent capitalization.
  ** Labels (and combobox values) use normal sentence capitalization (we had a mix of sentence capitalization, "title" capitalization, and "random" capitalization).
  ** Layout/section headers use "title" capitalization.
  ** Feature names are capitalized (Storage Group, Live TV, Job Queue, User Job)
  ** Proper capitalization of initialisms/acronyms.  (UPnP, PIN, etc.)
* Consistent naming of features (Live TV with a space--we had it both ways; Auto-Expire with a hyphen and capitalized as a feature name--we had it all sorts of ways)
* Miscellaneous grammar/typo corrections.
}}}

This change will have a large effect on translations.  The meanings of the strings don't really change, so translators will mostly just need to map their existing translations to the new source strings--and can take advantage of the opportunity to improve consistency in the translations, themselves, at the same time.

There are probably some places where I missed fixes, but this should be a good start.


git-svn-id: http://svn.mythtv.org/svn/trunk@25039 7dbf422c-18fa-0310-86e9-fd20926502f2