Fix UI hang at startup on a misconfigured system (generally showing a whitish, pre-themed screen--potentially with some text in tiny fonts at the top left).
Iff the DEFAULT_UI_THEME and FALLBACK_UI_THEME do not exist, MythUIHelper's m_themepathname gets changed from the empty string to "/".  This resulted in the MythFontManager's scanning the entire filesystem looking for font files (and, generally, falling into an infinitely deep well in /proc).  This change checks for "/" as well as empty directories before scanning.  I'll also add code to trunk to limit the number of directories the MythFontManager will scan, but this fixes the issue in a -fixes-compatible way.


git-svn-id: http://svn.mythtv.org/svn/trunk@25050 7dbf422c-18fa-0310-86e9-fd20926502f2