Add an explicit recursion limit to MythFontManager's directory scan in case things don't work as expected (as with [25050]).
The current limit is 100 directories per "user" (i.e. per theme)--which is overkill, but in testing took less than 40ms with heavily populated directories on my not-spectacular dev system.  In theory, though, we'll never actually hit the limit, so--presuming we scan the directories we should--scan time will be identical to before this patch (and much lower than the 40ms).


git-svn-id: http://svn.mythtv.org/svn/trunk@25052 7dbf422c-18fa-0310-86e9-fd20926502f2