Pause the decoder until we have finished seeking for a saved bookmark.
This adds an option to start the decoder thread paused, uses it when
there is a bookmark saved and then seeks to the bookmark before
unpausing the decoder.

It's worth noting that Jean-Yves is looking to move the audio device
creation out of AvFormatDecoder, which would probably have negated the
audio side effects, but the change is useful and valid regardless as it
prevents unnecessary decoding before the seek.

Closes #8543


git-svn-id: http://svn.mythtv.org/svn/trunk@25054 7dbf422c-18fa-0310-86e9-fd20926502f2