Add an example clause to the Custom Rule editor that shows how to create a rule that re-records HDTV versions of shows that were previously recorded in SDTV.
As mentioned in the example clause title, duplicate matching should be disabled for the custom rule (however, it will only match episodes for which you have recorded an SDTV version and have not yet recorded an HDTV version).  The clause works great with "Match an exact title," but can be used with virtually any other clause (or, in theory, alone).  You may want to lower the priority of the Re-record in HDTV rules so they don't bump your first-time-recording rules.

This change does not attempt to fix the issues with the Custom Rule editor usability (which is still awaiting a multiline edit widget).

Thanks to Taco Mel on the mailing lists for coming up with the SQL and to Johnny Walker for creating the patch.


git-svn-id: http://svn.mythtv.org/svn/trunk@25056 7dbf422c-18fa-0310-86e9-fd20926502f2