Fixes #8494. Cleanup of ProgList code.
I had tried to do a minimal fix of a large memory leak found in this class during the ProgramInfo refactor, but a high degree of bitrot meant that a bit of reorg was necessary to fix the memory leak safely. This does a full once over of the ProgList code fixing not just the memory leak but all problems seen. Note: The helper classes in proglist_helpers.{h,cpp} were not cleaned up, only moved to out.

git-svn-id: http://svn.mythtv.org/svn/trunk@25090 7dbf422c-18fa-0310-86e9-fd20926502f2