Remove the legacy non-MythUI OSD themes.
These themes do not work with the new OSD engine.  The XML is very different, so pretty much the only part that's reusable is the images and the positioning information.  Anyone wanting to use that information for a MythUI OSD theme can find it at http://svn.mythtv.org/trac/browser/trunk?rev=25116 .


git-svn-id: http://svn.mythtv.org/svn/trunk@25117 7dbf422c-18fa-0310-86e9-fd20926502f2