Don't offer an Auto Detect fill option in the OSD menu when it is not supported (i.e. when using hardware accelerated decoding).
I've also enhanced the code to allow pre-selection of the Auto Detect
option though this will not always work - as far as I can tell because
if you select Auto Detect and it is not required, the current fill mode
is reset to off. Hence it really needs a refactor to differentiate
between current auto fill mode and the desired mode.

Refs #8592.


git-svn-id: http://svn.mythtv.org/svn/trunk@25158 7dbf422c-18fa-0310-86e9-fd20926502f2