Major audio settings rewrite. The settings page is now fully dynamic and not populated with static entries. Capabilities are accessible according to what the audio device can actually support. PulseAudio server is now only suspended as required and only when audio needs to be played. User settings are now ignored if incompatible with selected audio device. Add 'Scan' button in the settings page that will fill the list of devices available. We do so in order to speed up mythfrontend startup time as scanning for all devices can take a few seconds.
Typical user scenario when mythfrontend is first started, is to go to the settings; Press the scan button and choose the audio device required. If entering a custom device, not appearing in the list, enter the name then press Scan. Capabilities will then be retrieved or the user will be prompted if it is invalid.

git-svn-id: http://svn.mythtv.org/svn/trunk@25164 7dbf422c-18fa-0310-86e9-fd20926502f2