Added a timeout to the icon grabbing for the icon importer.
Every one of the icon files we grab gives us a 301 Permanent Redirect, which we
then follow.  The problem is, we look up the DNS for every single request each
time we do a request, and have no local caching.  This causes slowness if the
configured DNS server is slow.

The best way to fix this is two-fold:
1) use faster resolvers (8.8.8.8 and 8.8.4.4 should work if nothing else -
   google's public resolvers)
2) do local caching within the application to reduce the DNS lookups
   significantly



git-svn-id: http://svn.mythtv.org/svn/trunk@25180 7dbf422c-18fa-0310-86e9-fd20926502f2