Rework DBConnection as a connection pool. Each opened cursor now receives its own connection, with a default pool size of 2. When the pool is exhausted, additional connections will be spawned as needed. This allows the database connection to be considered thread safe.
The pool size can be managed with the methods:
  DBConnection.setDefaultSize(size)  # classmethod
  DBConnection.resizePool(size)

The consequence of this change is that all cursors must be manually closed with:
  cursor.close()
Any cursor not closed will result in a leaked connection. This includes cursors lost during error handling. In order to facilitate this, connections and cursors now offer a context manager for the 'with' statement:
  db = MythDB()
  # connection serving as context manager
  with db as cursor:
    cursor.execute(...)
  # cursor serving as context manager
  with db.cursor() as cursor2:
    cursor2.execute(...)
    # nested use is allowed
    with db as cursor3:
      cursor3.execute(...)
In this syntax, the cursor will be automatically closed, and the connection released, as the interpreter exits the context, regardless of exception state.

Refs #8577.


git-svn-id: http://svn.mythtv.org/svn/trunk@25200 7dbf422c-18fa-0310-86e9-fd20926502f2