Improved rendering control in MythPainterD3D9.
Only clear the framebuffer if it is offscreen and only begin/end the
D3D9 scene and swap the buffers if the painter has overall control of
rendering. Otherwise the owner of the painter object is responsible for
beginning and ending the scene (to better manage performance) and to
swap the buffers.

These changes are required for upcoming Direct3D osd rendering fixes.


git-svn-id: http://svn.mythtv.org/svn/trunk@25223 7dbf422c-18fa-0310-86e9-fd20926502f2