Follow-up on r25210 ; expand custom override resolution screen matching.
Stop using the map/key search mechanisms; move key to a 64 bits integer as a 32 bit value couldn't contain the required range (frequency was limited to 64Hz).
As we are using floatting refresh rate, searching via a unique key was too inaccurate; so instead loop through the screen map, with a 10E-2 accuracy.

Look through the custom screens as follow:
1. search for exact match (width, height, rate)
2. search for resolution, ignoring rate
3. search for matching height and rate (or any rate if specified rate = 0)
4. search for 2x rate
5. search for 1x rate

As such, you can define a custom resolution for a given refresh rate.

git-svn-id: http://svn.mythtv.org/svn/trunk@25255 7dbf422c-18fa-0310-86e9-fd20926502f2