Fix a deadlock when the input changes and we're using the 'softblend' osd.
A stream change notified from the decoder thread triggers a re-
conditions this can cause a deadlock if the decoder has already been
paused by the player thread. So instead always re-initialise the OSD
from the main thread.

This fixes a deadlock with a clip provided by otto and while it is
treating the symptoms rather than the cause (fix to follow), it is safer
all round and will avoid a number of potential QObject thread related
issues.


git-svn-id: http://svn.mythtv.org/svn/trunk@25266 7dbf422c-18fa-0310-86e9-fd20926502f2