Add a GPL-ed ZIP file handling class by Angius Fabrizio so that we can natively extract ZIP archives without having to spawn out to external programs.  This is only the UnZip class portion.  The Zip side can be added later if we ever have the need.  This code will be used in an upcoming Theme Downloader patch to allow mythfrontend to automatically download and install new or updated themes from the Internet.  The source repository is located at the following:
    http://code.google.com/p/osdab/source/browse/trunk#trunk/OSDaB-Zip

The code has been imported as-is, no modifications have been made other
than to convert from dos to unix linefeeds.

The only dependency of this code is zlib which should already be
available since Qt requires it as well.

If you're wondering 'why zip instead of <insert your favorite compression
program that compresses better>', the reason is two-fold: 1) this code is
fairly simple and self-contained, and 2) the only current use of this
code will be for downloading theme packages and since themes are mainly
images we won't see a huge benefit in using a better compression
algorithm.

This modifies the binary API version, so make clean, etc..



git-svn-id: http://svn.mythtv.org/svn/trunk@25270 7dbf422c-18fa-0310-86e9-fd20926502f2