ThemeInfo changes to support the upcoming Theme Downloader.
- Allow ThemeInfo to directly download and parse a themeinfo.xml file
from a remote site.  The theme itself can't be run remotely, this
code is to allow the Theme Downloader UI to instantiate new ThemeInfo
objects without having to download remote themeinfo.xml files to temp
files on the local filesystem.

- Add 'downloadurl' and 'themesite' tags under the details section of
the theminfo.xml schema.  The downloadurl is the location of a the
theme package file which will be downloaded.  The themesite tag is
to hold an optional website if the author of the theme provides one.

NOTE: This updates the binary API version, so make clean, etc..



git-svn-id: http://svn.mythtv.org/svn/trunk@25275 7dbf422c-18fa-0310-86e9-fd20926502f2