Enable chapter handling in BD playback.
This enables the code committed in r25282:25283, adds some consistency
fixes between handling of Matroska and BD chapters and improves the log
output.

If anyone wants to further extend this to title support, feel free:) ...
or indeed add DVD title and chapter handling.

There is one outstanding bug related to seeking to the first chapter
(frame 0) which will be fixed shortly.


git-svn-id: http://svn.mythtv.org/svn/trunk@25284 7dbf422c-18fa-0310-86e9-fd20926502f2