Fixes #8641.
This makes sure mythui painting is disabled before video playback starts and that myth_system can not cause mythui painting to be re-enabled.

Note: In order to make sure the video intro screen is still shown I moved some of the tv_play initialization so that the screen is drawn earlier.

git-svn-id: http://svn.mythtv.org/svn/trunk@25319 7dbf422c-18fa-0310-86e9-fd20926502f2