Add two new Mashup grabbers from Doug Vaughan, Space and Technology.  I also have a few outstanding grabbers from Auric that I need to commit in the next few days when I find a little bit of time.  Also patch several of the existing grabbers with the following log:
> common_api.py
Add a function that allows grabbers to check if an item is already in the data base. This is particularly important when a grabber is forced to do additional Interent accesses to acquire all the needed MNV item data.
Added a function that checks if there are any treeview items in the data base for a specific
grabber. Some Mashup grabber's search option only returns results when then there are treeview
items in the database.
Made the creation of custom HTML page links more flexible so code did not need to be changed when new custom HTML pages were added.
Add the ability for parameters to be passed to a XSLT style sheet

> mashups_api.py
Add the ability for a Mashup to pass variables to a XSLT style sheet

> bbciplayer.html, bliptv.html, cinemarv.html, thewb.html, tributeca.html
Changed the position of the Flash control bar as it was partially hidden
Changed the colors of the Flash control bar on some of the custom HTML pages

> natureMashup.xml
Add the "DiveFilm HD Videos" source to the Nature grabber

> trailersMashup.xml
Removed the source "HD Movie Trailers Online" due to the RSS is no longer maintained

> cinemarvRSS_item.xsl, traileraddictRSS_item.xsl, tributecaHTML_item.xsl, cinemarv_api.py, traileraddicts_api.py, tributeca_api.py
The Trailers grabber now checks the local data base for existing item data rather that executing multiple Internet accesses. This greately reduces the processing times of this grabber.

> rev3_api.py
Fixed missing shows from the creation of the user default preference due to Web site changes
Fixed two incorrect variable names in debug messages

> bliptvXSL_api.py
Fixed a typo which caused the item to occasionally miss a download link

> nature.py
Toggle the search option off if there are NO "Nature" items in the "internetcontentarticles" table



git-svn-id: http://svn.mythtv.org/svn/trunk@25332 7dbf422c-18fa-0310-86e9-fd20926502f2