Backport of [25363] to 0.23-fixes.
Fixed a regression in handling the "delete, but allow re-record"
functionality.  When the playbackbox was reworked to make the GUI more
responsive, the delete action was made asynchronous.  If the program
was still being recorded when the user chose to delete but allow
re-record,, this resulted in the history getting rewritten by the
backend when the recording was stopped.  This change fixed the problem
by passing the desire to allow re-record in the DELETE_RECORDING
command so the backend can fix up the history after the program has
been stopped.

Note: this changes the ABI and protocol versions, so it's probably a
good idea to make clean before compiling.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-23-fixes@25364 7dbf422c-18fa-0310-86e9-fd20926502f2