Improve efficiency of mythlink.pl when called with --chanid and --starttime.
On my Athlon II 240 (3GHz, 4GB RAM) with 1360 recordings, this changes the time it takes for mythlink.pl to create a link for a single recording from about 6 seconds to 0.1 seconds.

Note for 0.23-fixes users:  This version of mythlink.pl will work with 0.23-fixes.  However, I don't plan to commit the change to 0.23-fixes since it's not fixing a bug.  So, if you use mythlink.pl with User Jobs or MythTV System Events, feel free to download the updated copy from trunk and use it.  See, also, http://www.mythtv.org/wiki/Mythlink.pl for more information on using mythlink.pl with User Jobs or MythTV System Events.


git-svn-id: http://svn.mythtv.org/svn/trunk@25386 7dbf422c-18fa-0310-86e9-fd20926502f2