Loosen restriction on directory permissions for directories created by mythlink.pl.
Previously, it was impossible to have mythlink.pl create a directory with group write permissions.  This change allows directories to be created with up to 0775 permissions (but permissions will be modified by umask, so may be more restrictive if desired).  Note that this change only affects directories created by mythlink.pl and will not affect already-existing directories (and has no effect on symlinks).



git-svn-id: http://svn.mythtv.org/svn/trunk@25391 7dbf422c-18fa-0310-86e9-fd20926502f2