Remove some old code that was used for an out-of-sequence database update.
In the 0.20-fixes days, we needed a database update to convert Zap2It Labs programid format to Schedules Direct format.  This was done in both 0.20-fixes and trunk with an out-of-sequence database update using a setting to determine if the update had been applied, and the code was kept in place to allow interoperability regardless of provider.

Since we no longer need the interoperability, this change simply moves the update to an official "in-sequence" update in trunk (though it has probably already run on everyone's system, so won't do anything), and removes the support code to simplify the code a bit.


git-svn-id: http://svn.mythtv.org/svn/trunk@25417 7dbf422c-18fa-0310-86e9-fd20926502f2