Modify UpdateDBVersionNumber() to accept the current dbver and update its value, just like performActualUpdate().
This will prevent our forgetting to manually update dbver after calling UpdateDBVersionNumber() for updates that don't use performActualUpdate() (which results in the upgrade sequence stopping at the point of the forgotten update, and MythTV unknowingly operating on an out-of-date schema version until the application is restarted).  The patch also adds "Upgrading to MythTV schema version xxxx" messages for all the "invisible" updates that use UpdateDBVersionNumber(), to ensure that there are no "skipped" version numbers in the log output.


git-svn-id: http://svn.mythtv.org/svn/trunk@25418 7dbf422c-18fa-0310-86e9-fd20926502f2