Do a quick check of database tables on master backend and mythtv-setup startup.
This change causes the master mythbackend and mythtv-setup to check the database tables at application startup.  In the event one or more tables is not OK (crashed or errored), it will attempt to repair the tables.  If the repair succeeds, startup will continue.  Otherwise, the application will exit with a message indicating that the user should repair the database tables.

The quick check should pick up crashed tables caused by improperly closed data files.  Some classes of data errors won't be detected (checksum errors or incorrect links), but this should handle the majority of issues caused by crashed/killed mysqld server and full filesystems.  Note, also, that clients (mythfrontend, mythavtest, ...) and remote backends will not check database tables, since it is not unusual for them to be started and restarted many times while the master backend continues to run.


git-svn-id: http://svn.mythtv.org/svn/trunk@25426 7dbf422c-18fa-0310-86e9-fd20926502f2