Update the OSD code to allow setting the timeouts for individual OSD elements to either a short, medium or long value.
These values are currently implemented at 3, 5 and 10 seconds but can be
modified. For consistency with the current implementation I've set
everything to timeout using the medium setting of 5s.

To complete the exercise, it should just need someone to change the
timeouts to the shorter or longer versions as appropriate and
agree/implement a method of customising them - either through the
settings page or the theme.


git-svn-id: http://svn.mythtv.org/svn/trunk@25465 7dbf422c-18fa-0310-86e9-fd20926502f2