MythMusic: Add the infrastructure that allows the music player to play from sources other than local files.
The idea is to eventually allow the music player to play files in storage
groups via a remote backend or from the internet via http or from Ice/Shoutcast
streams etc.

The DecoderHandler is responsible for starting the decoder via DecoderIOFactorys.
It operates on a playlist, either created with a single file, by loading a .pls
or downloading it, and for each entry creates an appropriate DecoderIOFactory.
The creator is simply an intermediate class that translates the next URL in the
playlist to a QIODevice. Ie. the DecoderIOFactoryFile just returns a QFile. It
would be possible to add a DecoderIOFactoryShoutcast which would return a
QTcpSocket subclass where reads do the necessary translation of the shoutcast
stream.

Most of this is taken from an old patch by Eskil Heyn Olsen to add Shoutcast
support to MythMusic. I've updated it for QT4, programming style as well as
making the necessary changes so it works with current trunk.  



 


 

git-svn-id: http://svn.mythtv.org/svn/trunk@25473 7dbf422c-18fa-0310-86e9-fd20926502f2