MythMusic: Add shoutcast playback to the music player.
This is still a work in progress but adds everything needed by the player to
download and parse the streams. There is no UI yet, that will come when we
switch MythMusic to use MythUI.  

Some of this is based on an old patch by Eskil Heyn Olsen although a lot has
been changed to make better use of the Qt4 networking stuff.


git-svn-id: http://svn.mythtv.org/svn/trunk@25524 7dbf422c-18fa-0310-86e9-fd20926502f2