Make the actual process of changing the channel part of the signal monitor.  Refs #6719
For LiveTV this allows feedback to be given to the user, instead of leaving
them in limbo until the channel change is complete.  Also allows for
situations where channel changes can take a very long time (e.g. using IR
blaster).

This does require the channel change process to be moved into it's own
thread, and is therefore a somewhat major change.


git-svn-id: http://svn.mythtv.org/svn/trunk@25543 7dbf422c-18fa-0310-86e9-fd20926502f2