Create a queue of player events to ensure that the video playback loop runs continuously.
The new playback event mechanism introduced in r25538 does not work well
when the framerate is at or near the refresh rate or the system is
otherwise busy. By maintaining a queue of playback events, we ensure the
next loop iteration is triggered as soon as possible while continuing to
process system events.

As commented in the code, if we start to lose events, playback will fail
but I'm not aware of situtations under which Qt drops or loses events.


git-svn-id: http://svn.mythtv.org/svn/trunk@25546 7dbf422c-18fa-0310-86e9-fd20926502f2