Add support for the VideoDecodeAcceleration framework on OS X.
This enables video acceleration on certain hardware when running OS X 10.6.3 and above (see http://developer.apple.com/mac/library/technotes/tn2010/tn2267.html)

There are 2 outstanding issues:-

- there is no interlaced and repeat picture detection (and hence no deinterlacing). This will require some additional parsing or perhaps a tie in into the existing H264 parser.
- some H264 broadcast streams are broken. So far the only problem seems to be BBC HD.

Implementation wise, the VDA library is loaded dynamically in order to simplify package management (the framework is only available on newer versions of OS X) and the additional privatedecoder_vda_defs file replicates the required elements of the framework headers as the headers are currently private.

There is a clear performance improvement, though nothing approaching the gains seen with, for example, VDPAU on the same hardware - no doubt due to the additional overhead of transferring the frames back into main memory (and back to GPU memory again).

The interaction with the Quartz video renderer is untested.

Refs #8621.


git-svn-id: http://svn.mythtv.org/svn/trunk@25571 7dbf422c-18fa-0310-86e9-fd20926502f2