Disable some pixel format request code in PrivateDecoderVDA.
kCVPixelBufferIOSurfacePropertiesKey is OS X 10.6 only and the property is assumed by default.
The format of the returned buffer should be either BGRA or yuv2 anyway, both of which we can deal with. We will however need to request specific formats if we are going to pass the decoded frames directly to the OpenGL renderer.


git-svn-id: http://svn.mythtv.org/svn/trunk@25618 7dbf422c-18fa-0310-86e9-fd20926502f2