Fixes #8736. This makes the multirec behaviour consistent whether you enter Recording Options or not.
This always creates one virtual card for DVB/HDHR unless the user explicitly changes this in the Recording Options screen.

The number of cards created is now controlled kDefaultMultirecCount, which is set to 2. If we want to change it to 1 or 3 or some other number in the future it is now just a policy decision as it is trivial to do.



git-svn-id: http://svn.mythtv.org/svn/trunk@25629 7dbf422c-18fa-0310-86e9-fd20926502f2