Fix issues with DiSEqC caused by [25543].
Creating the LiveTVRingBuffer before the channel change operation means that
the channelbase::sourceid may not be up to date when multiple sourceids are
associated with a single input.  Therefore a call to CheckChannel is needed
to determine that information.

Also, when initializing the inputs in ChannelBase, retrieve the "default
input" from the DB in case the capture device is opened "on demand".

Both of these issues could result in incorrect program information being
displayed when changing from one channel to another in LiveTV, as well as
tuning failures due to DB lookup problems.

Thanks to Gus Bourg for discovering and diagnosing the problem with DiSEqC.

Thanks to Benny Sjöstrand for discovering the issue with opening devices "on
demand", and fixing it.

Thanks to Mark Buechler for his help in analyzing log files, and diagnosing
the problems.

Fixes #8742
Fixes #6719


git-svn-id: http://svn.mythtv.org/svn/trunk@25642 7dbf422c-18fa-0310-86e9-fd20926502f2