Added commercial detection speedups.  This works primarily by using a scaled down version of the recording to detect commercials (1/4 width, 1/4 height), and by doing blank frame detection using less of the frame.
Preliminary testing has shown that the commercial breaks found are as accurate
if not more accurate with this enabled as with it disabled, with occasional
missed commercials.  This seems no worse than without it, and commercial
detection is not a simple thing anyways.

There is a new setting in the TV -> General (Jobs) screen to allow the user to
enable these speedups, and the default is to have the speedups disabled.

Closes #6824



git-svn-id: http://svn.mythtv.org/svn/trunk@25657 7dbf422c-18fa-0310-86e9-fd20926502f2