Adds "Reduce to fit" scaling option to mythgallery scaling.  The scaling options now include:  - scale to fit  - scale to fill  - reduce to fit.
Reduce to fit is the same as scale to fit with the exception that it will never
scale up a small picture.

Scaling options can be cycled while viewing a picture by hitting the 
"FULLSCREEN" command (by default W key)

Applied a reworked version of the patch from #5897.

Thanks to Boleslaw Ciesielski <bolek-mythtv@curl.com> for the patch.

closes #5897



git-svn-id: http://svn.mythtv.org/svn/trunk@25659 7dbf422c-18fa-0310-86e9-fd20926502f2