Adds support for the various raw photo formats as decoded by dcraw.  As dcraw is only distributed as a binary, this actually creates a Qt plugin for mythgallery that runs dcraw externally.
There are likely to be some RAW photo extensions that dcraw supports that got
missed by the original author, and will need to be added.

Also, I decided to leave the icc-profile as a configure-time setting for the
moment as there's a push to be redoing all settings soon anyways, and we might
as well redo it then.

Original patch from Chad Parry <spam@chad.parry.org>
Closes #7745



git-svn-id: http://svn.mythtv.org/svn/trunk@25667 7dbf422c-18fa-0310-86e9-fd20926502f2