Revisit the automatic caption track selection.
The recent OSD menu changes updated the way captions and subtitles are
presented to the user. The VBI format is no longer used to restrict the
selection and we decide based upon the available scanned streams (with
the exception of the Teletext Menu). This makes it simple to access
captions from videos that were captured outside of your own 'region'.

This change applies the same logic to the automatic caption selection,
with only the NUV teletext caption type restricted to those who have set
their VBI format to PAL.

Also helps with debugging!


git-svn-id: http://svn.mythtv.org/svn/trunk@25697 7dbf422c-18fa-0310-86e9-fd20926502f2