Improved mutli-threaded approach to subtitle initialisation.
Firstly, improve the teletext viewer setup from r25696 and ensure the
viewer is always created by the player thread. This was causing problems
when subtitles are on by default and are initialised early by the
decoder thread.

Secondly, apply a similar fix to the TracksChanged method to ensure
subtitles are enabled from the player thread.


git-svn-id: http://svn.mythtv.org/svn/trunk@25701 7dbf422c-18fa-0310-86e9-fd20926502f2