Pass a 'tvstate' string to the OSD.
Refactor StateToString to provide only debug versions of the state
strings, which don't require 'brief' versions or translations and a
missing BD string. Add a new StateToPublicString version to provide
translatable strings that can be safely printed by the UI. Add the tv
state string the to program information map passed to the OSD.

I've not added this to any themes - clearly it's up to themers to use at
their discretion.

Closes #3632


git-svn-id: http://svn.mythtv.org/svn/trunk@25705 7dbf422c-18fa-0310-86e9-fd20926502f2