Optimise teletext rendering.
The post OSD-merge teletext rendering used a libmythui object for each
character and for each character background. This meant the hardware
accelerated UI painters (VDPAU, OpenGL and Direct3D9) were possibly
processing over 2,000 elements for each frame rendered (hence up to
100,000 per second) - and there was a sizeable performance hit.

The new approach is to render to some intermidiary QImage buffers for
each double line and when complete, convert those to MythUIImage. This
is substantially faster, with a maximum of 13 objects to be rendered per
frame.


git-svn-id: http://svn.mythtv.org/svn/trunk@25713 7dbf422c-18fa-0310-86e9-fd20926502f2