Add a little fuzzyness to the main title selection in BD playback.
I have at least one disk where the main title is first but the length is
reported as being 5 seconds shorter than the audio commentary, standard
definition title. The main titles seem in general to be earlier in the
selection, so prefer the earlier title if it is only fractionally
shorter (set to around 30 seconds at the moment - may need tweaking).


git-svn-id: http://svn.mythtv.org/svn/trunk@25736 7dbf422c-18fa-0310-86e9-fd20926502f2