Don't use the deleteLater method in MythUIType.
The current libmythui usage model suggests this is unnecessay but it is
causing problems in the OSD. MythImages are allocated via the painter
classes but owned by their parent MythUIImage objects. When the painter
is deleted immediately and the MythUIImage is not deleted until the
event loop has finished, the painter pointer held by MythImage is no
longer valid.

This is the simplest and most obvious solution but if there are
unforeseen consequences, there is a less elegant solution.

Closes #8586. Refs r25740


git-svn-id: http://svn.mythtv.org/svn/trunk@25741 7dbf422c-18fa-0310-86e9-fd20926502f2