Keep a record of the original number of audio channels for each audio track.
If the number of AC3 channels exceeds the number of available speakers
and we are not using passthrough, the number of channels recorded in
AvCodecContext is overwritten. This leads to a couple of presentational
problems in the UI - the number of channels shown for a given audio
track can mysteriously change and the number of channels shown in the
OSD audio information is wrong.

We now store the original channel count in the StreamInfo, pass this to
the AudioPlayer and AudioInfo classes and use the original count for
items presented to the user.


git-svn-id: http://svn.mythtv.org/svn/trunk@25755 7dbf422c-18fa-0310-86e9-fd20926502f2