Add support for animated images in MythUIImage.  The main use for this is displaying animated gifs loaded directly from http:// sources such as the animated radar images that MythWeather will be displaying.
The images are not saved in the image cache currently.  This includes
the on-disk and in-memory caches.  If we are going to cache the individual
images, then we also need to cache the delay values between images, so the
issue of how to store that info needs to be resolved before we can cache
the images themselves.

References #8625.  I'm going to leave the ticket open for now and change
the milestone.  The basic support added in this patch is enough for 0.24
and the caching support will be added post 0.24.



git-svn-id: http://svn.mythtv.org/svn/trunk@25801 7dbf422c-18fa-0310-86e9-fd20926502f2