remove "Cross source EIT" setting
it is unneeded since the supported cases can be handled by prefering the
current sourceid. This still works on diseqc settings since the current
source will not have that channel if it is on a different source.

Setups with multiple channels with the same networkid, transportid,
serviceid triple might "break" since a different channel gets the
program data now. This is unsupported and it was pure luck that it
"worked" before. Refs #7672

git-svn-id: http://svn.mythtv.org/svn/trunk@25813 7dbf422c-18fa-0310-86e9-fd20926502f2