MythVideo/MythTV:  Move all of the behind the scenes scanning, metadata, and DB manipulation of MythVideo into MythTV.
Takes the parental controls, video metadata, videometadata DB access, metadata class, metadatalistmanager, scanning, and global variable defaults and creates a new core library called "libmythmetadata."

Initially, libmythmetadata only exists to serve mythvideo.  As MythVideo is destined for core MythTV, this is less of a stretch than one might imagine.  This also opens the door for the backend and uPnP to take advantage of MythVideo's scanning and metadata classes.  It should be a fairly minor task to add automatic scanning and metadata updates to the backend, allow the uPnP server to access the full breadth of metadata available (while eliminating its own scanner).

A couple of tasks remain outstanding:

1)  We need a core DB update to add videometadata for all Myth systems, not just those with MythVideo installed.

once this is complete, we can proceed to:

2) Move the UI portions of MythVideo into the core, eliminating it as a plugin.

3) Add scanner and metadata downloader classes (both already complete) to the backend, run them on a regular schedule.  This will require a management interface for media that is not found during a scan-- we need to notify the user and get their okay before flushing it from the DB.

4) Have the uPnP server use the libmythmetadata/metadata.cpp classes to get lists of all the content available, their artwork, etc.

5) ???

6) Profit!

This bumps the lib version, you will need to make clean and recompile plugins.


git-svn-id: http://svn.mythtv.org/svn/trunk@25895 7dbf422c-18fa-0310-86e9-fd20926502f2