Extend VideoBuffers and VideoFrame to support frame formats other than YUV420P.
This isn't complete, and hence there are no functional changes, but
begins to remove some of the hard-coded assumptions of a YUV420P frame
format. Other formats are becoming increasingly relevant for VDA and
CrystalHD hardware decoding and the Direct3D, OpenGL and XVideo
renderers can all be modified to accept other formats and hence avoid
relatively costly conversions.


git-svn-id: http://svn.mythtv.org/svn/trunk@25936 7dbf422c-18fa-0310-86e9-fd20926502f2